requirements.txt
antropy==0.1.4
autoreject==0.3.1
chart_studio==1.1.0
click==8.1.3
h5io==0.1.7
kaleido==0.2.1
matplotlib==3.5.3
mne==1.1.1
nolds==0.5.2
numpy==1.23.2
omegaconf==2.2.3
openpyxl==3.0.10
pandas==1.4.4
plotly==5.10.0
PyQt5==5.15.7
python-dotenv==0.21.0
scikit_learn==1.1.3
scipy==1.9.1
setuptools==58.1.0
torch==1.13.0
torchvision==0.14.0
